Maintainability wave: S2094/S1939/S7130/S1905/S2933/S3247/S6561; prune Sonar export
- Add FluxorPayloadVersion const to marker actions/record (S2094); drop redundant IListViewModel from IHasFetchError* view models (S1939).
- Remove redundant SqliteErrorCode int casts (S1905); use First() when melody dictionary is non-empty (S7130).
- MusicSectionSelection stateChangeHandler readonly; IsCategorySelectable readonly (S2933).
- DayColorConverter and DrawableToBitmap pattern matching (S3247); UtcNow for permission cache and resume timing deltas (S6561).
- Sonar export total 552 -> 530.

+++ src/Bible.Alarm.Shared/Services/Media/Helpers/SectionFetcherHelpers/SectionFetcherSqliteExceptionHelper.cs
@@ -16,7 +16,7 @@ public static class SectionFetcherSqliteExceptionHelper
         {
             if (e is SqliteException sqliteEx)
             {
-                var code = (int)sqliteEx.SqliteErrorCode;
+                var code = sqliteEx.SqliteErrorCode;
                 if (code is 5 or 6)
                 {
                     return true;
@@ -31,7 +31,7 @@ public static class SectionFetcherSqliteExceptionHelper
     {
         for (var e = ex; e != null; e = e.InnerException)
         {
-            if (e is SqliteException sqliteEx && (int)sqliteEx.SqliteErrorCode == 19)
+            if (e is SqliteException sqliteEx && sqliteEx.SqliteErrorCode == 19)
             {
                 return true;
             }
